Clean up compiler warnings  (#37)
* clean up compiler warnings

* small change in ui_config

+++ src/monsters/fall_boss.cpp
@@ -6,10 +6,10 @@ const std::string WALK_SPRITE = "bear/bear_walk.png";
 const std::string RUN_SPRITE = "NA";
 const std::string ATTACK_SPRITE = "bear/bear_attack.png";
 const std::string DEATH_SPRITE = "bear/bear_death.png";
-const size_t WALK_FRAMES = 6.f;
-const size_t RUN_FRAMES = 0.f;
-const size_t ATTACK_FRAMES = 7.f;
-const size_t DEATH_FRAMES = 6.f;
+const int WALK_FRAMES = 6;
+const int RUN_FRAMES = 0;
+const int ATTACK_FRAMES = 7;
+const int DEATH_FRAMES = 6;
 
 entt::entity FallBoss::createFallBossEntt()
 {
@@ -58,10 +58,10 @@ entt::entity FallBoss::createFallBossEntt()
     monster.death_sprite = DEATH_SPRITE;
 
     Animate& animate = registry.emplace<Animate>(entity);
-    animate.frame = 0.f;
-    animate.state = 0.f;
+    animate.frame = 0;
+    animate.state = 0;
     animate.frame_num = WALK_FRAMES;
-    animate.state_num = 1.f;
+    animate.state_num = 1;
 
     registry.emplace<FallBoss>(entity);
     registry.emplace<HitReaction>(entity);
