+++ src/monsters/winter_boss.cpp
@@ -6,10 +6,10 @@ const std::string WALK_SPRITE = "winter/winter_walk.png";
 const std::string RUN_SPRITE = "winter/winter_run.png";
 const std::string ATTACK_SPRITE = "winter/winter_attack.png";
 const std::string DEATH_SPRITE = "winter/winter_death.png";
-const size_t WALK_FRAMES = 4.f;
-const size_t RUN_FRAMES = 0.f;
-const size_t ATTACK_FRAMES = 0.f;
-const size_t DEATH_FRAMES = 0.f;
+const int WALK_FRAMES = 4;
+const int RUN_FRAMES = 0;
+const int ATTACK_FRAMES = 0;
+const int DEATH_FRAMES = 0;
 
 entt::entity WinterBoss::createWinterBossEntt()
 {
@@ -61,10 +61,10 @@ entt::entity WinterBoss::createWinterBossEntt()
     monster.death_sprite = DEATH_SPRITE;
 
     Animate& animate = registry.emplace<Animate>(entity);
-    animate.frame = 0.f;
-    animate.state = 0.f;
+    animate.frame = 0;
+    animate.state = 0;
     animate.frame_num = WALK_FRAMES;
-    animate.state_num = 1.f;
+    animate.state_num = 1;
 
     registry.emplace<WinterBoss>(entity);
     registry.emplace<HitReaction>(entity);
